feat: Complete real-time system, two-phase commit, and infrastructure improvements
Real-time WebSocket System:
- Implement production-ready WebSocket with automatic reconnection
- Fix 60-second timeout using WebSocket PING/PONG frames
- Add Redis pub/sub integration for block.indexed events
- Create frontend hooks: useWebSocket, useAllBlockEvents
- Add server-side filtering by chain subscription
- Implement panic recovery in all goroutines for stability

Two-Phase Commit Pattern:
- Update all 8 entities to use staging → production tables
- Add InsertBlocksStaging, InsertTransactionsStaging, InsertBlockSummariesStaging
- Ensure atomic promotion via ClickHouse ALTER TABLE EXCHANGE
- Prevent partial data visibility during indexing

Dynamic Entity Routing:
- Add route_path field to EntityInfo response
- Implement backend entityRouteMapping for all entities
- Update frontend to build dynamic entityMap from backend
- Eliminate hardcoded route mismatches (dash vs underscore)

OpenAPI Specification:
- Remove 845 lines of duplicate schemas and responses (~30% reduction)
- Fix malformed YAML structure with proper sections
- Document all 35 endpoints (13 Admin + 22 Query)
- Consolidate 18 duplicate schema definitions

Code Quality:
- Apply go fmt to all files
- Fix all 47 golangci-lint issues (now 0 issues)
- Update 50+ mock implementations for new staging methods
- Add nolint comments for test helpers
- Remove unused imports and variables

Status: All tests passing, Linter clean, Production ready

+++ pkg/rpc/block.go
@@ -82,7 +82,7 @@ func (bbh *BlockByHeight) ToBlockModel() *indexer.Block {
 		Time:            time.UnixMicro(bbh.BlockHeader.Time),
 		LastBlockHash:   bbh.BlockHeader.LastBlockHash,
 		ProposerAddress: bbh.BlockHeader.ProposerAddress,
-		Size:            bbh.Meta.Size,
+		Size:            int32(bbh.Meta.Size),
 	}
 }
